+++ scripts/formapplications/item-pile-inventory.js
@@ -233,8 +233,6 @@ export class ItemPileInventory extends FormApplication {
 
         data.systemHasCurrencies = !!API.CURRENCIES.length;
 
-        console.log(data.systemHasCurrencies)
-
         data.playerActors = this.playerActors;
         data.moreThanOneInspectableActors = data.playerActors.length > 1;
 
@@ -256,8 +254,6 @@ export class ItemPileInventory extends FormApplication {
         data.hasItems = data.items.length > 0;
         data.hasCurrencies = data?.currencies?.length;
         data.canInspectItems = pileData.canInspectItems || game.user.isGM;
-
-        data.hasThings = data?.hasItems && (data.systemHasCurrencies && data?.hasCurrencies);
 
         data.isEmpty = !data?.hasItems && !data?.hasCurrencies;